+++ CHANGELOG.md
@@ -13,3 +13,25 @@ Features and bug fixes
 
 Breaking changes
 - all breaking changes are in the API. See `guides/migration-5to6.md` for details.
+
+#### v6.1.0
+
+Features and bug fixes
+- enable inserting trace IDs into logs automatically (pino, winston)
+- new API function, `ao.getFormattedTraceId()` to get trace IDs using code.
+- new API function, `ao.sendMetric()` for sending custom metrics.
+- fix config file not found bug.
+
+#### v6.2.0
+
+Features
+- enable inserting trace IDs into logs automatically (bunyan)
+- `insertTraceIdsIntoLogs` options expanded to `false`, `true`, `'traced'`, `'sampledOnly'`, `'always'`
+
+#### v6.3.0
+
+Features and bug fixes
+- add `insertTraceIdsIntoMorgan` to enable appending `ao.traceId=...` format morgan's text output.
+- add `createTraceIdsToken`. Set to `'morgan'` to have the token `ao-auto-trace-id` token created. Use as `:ao-auto-trace-id` in morgan formats.
+- set config to service key that was used.
+- fix aws-sdk bad signature error on transaction retry.